crud angular material

+++ src/app/components/encuesta/encuesta-list/encuesta-list.component.ts
@@ -1,13 +1,21 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { Encuesta } from '../../../models/encuesta';
 import { ActivatedRoute, Route, Router, RouterLink } from '@angular/router';
 import { EncuestaService } from '../../../services/encuesta.service';
 import { NgFor } from '@angular/common';
+import { MaterialModule } from '../../../material.module';
+import { MatTableDataSource } from '@angular/material/table';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { EncuestaFormComponent } from '../encuesta-form/encuesta-form.component';
+import { MatDialog } from '@angular/material/dialog';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-encuesta-list',
   standalone: true,
-  imports: [NgFor, RouterLink],
+  imports: [NgFor, RouterLink, MaterialModule, FlexLayoutModule],
   templateUrl: './encuesta-list.component.html',
   styleUrl: './encuesta-list.component.css'
 })
@@ -15,7 +23,13 @@ export class EncuestaListComponent {
 
   encuestas: Encuesta[]= [];
 
-  constructor(private route: ActivatedRoute, private router: Router, private encuestaService: EncuestaService){}
+  displayedColumns = ['titulo','accion'];
+  dataSource = new MatTableDataSource<Encuesta>();
+
+  @ViewChild(MatSort) ordenamiento?: MatSort;
+  @ViewChild(MatPaginator) paginador?: MatPaginator;
+
+  constructor(private encuestaService: EncuestaService, private dialog: MatDialog, private router: Router){}
 
   ngOnInit():void{
     this.listarEncuesta();
@@ -23,18 +37,61 @@ export class EncuestaListComponent {
 
   listarEncuesta():void{
     this.encuestaService.obtenerTodasLasEncuestas().subscribe( encuestas =>{
-      this.encuestas = encuestas;
+      this.dataSource.data = encuestas;
     })
+  }
+
+  ngAfterViewInit() {
+    this.dataSource.sort = this.ordenamiento ?? null;
+    this.dataSource.paginator = this.paginador ?? null;
+  }
+
+  abrirDialog(operation: string, encuestaId?: number): void {
+    const dialogRef = this.dialog.open(EncuestaFormComponent, {
+      width: '550px',
+      data: { operation, encuestaId },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      this.listarEncuesta();
+    });
   }
 
   eliminarEncuesta(encuestaId: number):void{
-    this.encuestaService.eliminarEncuesta(encuestaId).subscribe(()=>{
-      this.listarEncuesta();
-    })
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#716add",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.encuestaService.eliminarEncuesta(encuestaId).subscribe(()=>{
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          }).then(() => {
+            this.listarEncuesta();
+          })
+        });
+      }
+    });
   }
 
-  verPreguntas(encuestaId: number):void{
-    this.router.navigate(['/encuestas', encuestaId, 'preguntas']);
+  buscar(event: KeyboardEvent):void{
+    const inputBuscar = event.target as HTMLInputElement;
+    this.dataSource.filter = inputBuscar.value.trim().toLowerCase();
+  }
+
+  verPreguntas(encuestaId: number, titulo: string):void{
+    this.router.navigate(['/encuestas', encuestaId, 'preguntas'], { state: { titulo: titulo } });
+  }
+
+  verDetallePregunta(encuestaId: number):void{
+    this.router.navigate(['/encuestas', encuestaId]);
   }
 
 }
